+++ runner/anvil_runner/__init__.py
@@ -1,3 +1,3 @@
 """Anvil privileged runner."""
 
-__version__ = "0.11.0"
+__version__ = "0.12.0"